feat: Event 수정 기능 구현
