+++ service/main-server/src/main/java/org/codeNbug/mainserver/domain/manager/entity/Event.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 
 @Entity
 @Table(name = "event")
-@Getter
+@Data
 @NoArgsConstructor
 @AllArgsConstructor
 @EntityListeners(AuditingEntityListener.class)
