+++ service/main-server/src/main/java/org/codeNbug/mainserver/domain/seat/repository/SeatRepository.java
@@ -2,6 +2,7 @@ package org.codeNbug.mainserver.domain.seat.repository;
 
 import java.util.List;
 
+import org.codeNbug.mainserver.domain.manager.entity.Event;
 import org.codeNbug.mainserver.domain.seat.entity.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,4 +11,6 @@ import org.springframework.data.repository.query.Param;
 public interface SeatRepository extends JpaRepository<Seat, Long> {
 	@Query("SELECT s FROM Seat s JOIN FETCH s.gradeId WHERE s.layout.id = :layoutId ORDER BY s.location ASC")
 	List<Seat> findAllByLayoutIdWithGrade(@Param("layoutId") Long layoutId);
+
+	List<Seat> findByEvent(Event event);
 }
