+++ service/main-server/src/main/resources/application.yml
@@ -16,7 +16,7 @@ spring:
     password: ${db.password}
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
     show-sql: false
     properties:
       hibernate:
